requirements.txt
discord.py>=2.0.0
yt-dlp>=2021.12.1
PyNaCl>=1.4.0
python-dotenv>=0.19.0
asyncio>=3.4.3